backend/utils/logger.py
# backend/utils/logger.py
import logging
import os
from dotenv import load_dotenv

# Load environment variables from .env file located at the project root
dotenv_path = os.path.join(os.path.dirname(__file__), '..', '..', '.env')
load_dotenv(dotenv_path=dotenv_path)

def setup_logger():
    """Configures and returns a logger instance."""
    log_level_name = os.getenv('LOG_LEVEL', 'INFO').upper()
    log_level = getattr(logging, log_level_name, logging.INFO)

    logging.basicConfig(
        level=log_level,
        format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
        datefmt='%Y-%m-%d %H:%M:%S'
    )
    logger = logging.getLogger(__name__)
    logger.info(f"Logger initialized with level: {log_level_name}")
    return logger

logger = setup_logger()