requirements.txt
Flask>=2.0
python-dotenv>=0.15
requests>=2.25
gunicorn>=20.1 # For production deployment